tfdbg CLI: Analyzer handlers: list_tensors & node_info
The "analyzer" is one of the two major components of the debugger, the other component being the to-be-implemented "stepper", which performs actions such as "continue to", "step" and "breakpoints" on a TF graph. The analyzer performs post hoc analysis of dumped tensor and graph data from debugged Session.run() calls.

This CL implements the first two command handlers for the CLI backend of the debugger analyzer: list_tensors & node_info.

Future CLs will check in other command handlers for the analyzer, including but not necessarily limited to: list_nodes, print_tensor, node_inputs and node_outputs.
Change: 133712226