Add initial version of MutableDenseHashTable that is implemented using tensors to store the key and value buckets. The implementation was inspired by the C++ dense_hash_map from https://github.com/sparsehash/sparsehash
Compared to the existing MutableHashTable it can be more memory efficient
when using small keys and values because it avoids the padding added by the
memory allocator and does not need to store pointers in the buckets.
For checkpointed tables it also avoids creating additional tensors during
checkpoint and restore operations.

Note that this is an initial implementation that is missing many features.
They will be added shortly.

Also update LookupInterface and existing implementations to pass down
OpKernelContext, which is required for MutableDenseHashTable.
Change: 135746705