go: Introduce a Scope for graph construction.
The Scope encapsulates common properties of operations being added to the
graph. The plan is for all generated wrapper functions for ops to consume a
Scope as their first argument. This is similar to the Scope class in the C++
API.

The Scope object defined here is far from complete, but I wanted to get this
out so that I can start on code generation for the wrapper functions for ops.

While at it, also defined named constants for DataType.

Another step in the journey towards #10
Change: 135882730