go: Use the generated op function wrappers in the example.
This commit makes the inception inference example use the graph construction
API, and in particular the generated functions for TensorFlow ops, to
normalize an image before sending for inference through the Inception model.

Besides demonstrating the use of the graph construction API, this also
makes the example usable on arbitrarily sized images. (Prior to this change,
the input image had to be precisely 224x224 pixels).

A slight variation of this would have involved the use of the ReadFile and
DecodeJpeg ops, however those are not usable in Go at the moment since
string valued tensors aren't supported yet.
Change: 137099275