Summary ops replace all illegal characters in their name with underscores.
Here's the context. The old summary ops (tf.scalar_summary, etc) took as their first argument a "tag", which was used to organize their output, and was totally independent of the TensorFlow naming system. This had some major disadvantages: it disconnected summary output from the graph (so it is not easy to map from summary data to the node that produced it), and since tags are required to be unique, it forced users to re-invent a hierarchical namespacing system to organize their summary data.

The new summary ops (tf.summary.scalar, etc) simply take a node name as the first argument. This means that the summaries will naturally leverage the tensorflow namespacing system and integrate nicely with name_scope, etc.

However, the old summary ops allowed any characters in tags, so replacing old summaries with new summaries would break some existing usage. For backwards compatibility, we replace illegal characters with underscores and log a warning.

This CL extends the replacement so instead of merely replacing spaces, it replaces all illegal characters.
Change: 137173533