tfdbg core: support using different tensor watches across DirectSession::Run calls
Prior to this CL, it is impossible to call DirectSession::Run multiple times with the same inputs and outputs, but with different debug_tensor_watches values. If you attempt to do that, the debug_tensor_watches value used during the first call will persist in later calls.

However, during debugging, it is sometimes desirable to change debug_tensor_watches across runs. For example, perhaps a client wants to watch no tensors in the first N training runs, but would like to watch some tensors in the (N+1)-th training run.

To achieve this, we do the following:
1) Let debug_tensor_watches generate a summary string;
2) Let this summary string be a part of the executor key in DirectSession.

Other minor changes:
* Replace "tfdb" with "tfdbg" as per the debugger's acronym change.
Change: 137176222