Change quantize_graph in eightbit mode to remove FakeQuantWithMinMaxVars nodes and use the information provided by them to set the min/max values on quantization-related nodes.
In eightbit mode, also changed how constant weights are quantized - instead of
doing it as a step after the main recursion, do it during the main recursion.
This allows the float inputs to FakeQuantWithMinMaxVars to be excluded from
quantization.

In eightbit mode, maintain more state in the stack during recursion.

Also change quantize reshape registration to register always and not use
TF_CALL_xyz; this matches other quantized ops.
Change: 137877226