XLA open-sourcing: remove dependency on util/process/subprocess.h.
Add to TensorFlow a subset of the functionality of SubProcess to support the
open-sourceable parts of XLA and support TensorFlow's own testing.

On google3 let it use the internal SubProcess from google3.  In the open-source
world, use the TF implementation of the feature subset.

Modify XLA and some TF tests to use the TF SubProcess.
Change: 139114815