Scripts to produce a tarball containing the TensorFlow C-library and C API header file.
These scripts are intended to be run with every release to
produce libtensorflow.tar.gz for CPU and GPU on Linux and OS X
for x86_64 architecture machines.
(Eventually there will be other operating systems and architectures).

These binary releases are then intended to make use of other language
bindings (such as https://github.com/tensorflow/rust,
https://github.com/tensorflow/haskell,
https://godoc.org/github.com/tensorflow/tensorflow/tensorflow/go) easier
as the common case would be to download the binary C-library release and
avoid the need to build TensorFlow from source (and all the time and
external dependencies doing so entails).