Java: Add support for scalar Tensors with dtype = TF_STRING
This change adds support for string scalars used by
popular operations like DecodeJpeg. Vectors, matrices
and higher-dimensional string tensors are left as an
excercise for the future.

Note that the TF_STRING data type in TensorFlow corresponds to a byte[] in
Java (not a String) since it is used for arbitrary bytes, not just
strings of unicode characters.

One more step in the journey that is #5
Change: 141221096