Before the embedding projector loads, check if WebGL is enabled (this technique is taken straight from Mozilla docs: https://developer.mozilla.org/en-US/docs/Learn/WebGL/By_example/Detect_WebGL).
If it is not, show a modal dialog as well as send a google analytics logging event.

To test, go to chrome://settings and uncheck "enabled hardware acceleration".

Note 1: We don't provide instructions for enabling WebGL. It's device / browser dependent, and we don't want to be in the business of keeping it in sync with the times.
Note 2: The error is centered, which I don't really like, but it's how we do all events so I'm keeping it this way for consistency.
Change: 141804360